Print declarations
